+++ src/renderer/berryflow/src/contexts/BerryFlowContext.tsx
@@ -1,49 +1,14 @@
-import React, {
-  createContext,
-  ReactNode,
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-} from "react";
+import React, { ReactNode, useCallback, useEffect, useState } from "react";
 import type {
   BerryFlowTemplate,
   BerryFlowWorkflow,
   WorkflowExecution,
 } from "../types/workflow";
 import { mockTemplates } from "../mocks/templates";
-
-interface BerryFlowContextType {
-  templates: BerryFlowTemplate[];
-  customWorkflows: BerryFlowWorkflow[];
-  currentWorkflow: BerryFlowWorkflow | null;
-
-  isBuilderOpen: boolean | null;
-  selectedTemplate: BerryFlowTemplate | null;
-
-  currentExecution: WorkflowExecution | null;
-
-  openBuilder: (template?: BerryFlowTemplate) => void;
-  closeBuilder: () => void;
-  selectTemplate: (template: BerryFlowTemplate) => void;
-  saveWorkflow: (workflow: BerryFlowWorkflow) => void;
-  loadWorkflow: (workflowId: string) => void;
-  executeWorkflow: (workflow: BerryFlowWorkflow) => Promise<void>;
-
-  isLoading: boolean;
-}
-
-const BerryFlowContext = createContext<BerryFlowContextType | null>(null);
-
-export const useBerryFLow = (): BerryFlowContextType => {
-  const context = useContext(BerryFlowContext);
-
-  if (!context) {
-    throw new Error("useBerryFlow must be used within a BerryFlowProvidewr");
-  }
-
-  return context;
-};
+import {
+  BerryFlowContext,
+  BerryFlowContextType,
+} from "../hooks/useBerryFlow.ts";
 
 export const BerryFlowProvider: React.FC<{ children: ReactNode }> = ({
   children,
